integrate PCA mux

+++ vehicle/mkvi/software/bms/tasks/tasks.h
@@ -21,7 +21,7 @@
  *   Number of PEC errors that occurred when communicating with the ADBMS1818
  *   0 if no errors
  */
-int voltage_task(uint16_t *pack_voltage, uint8_t *ov, uint8_t *uv);
+int voltage_task(uint16_t *pack_voltage, uint32_t *ov, uint32_t *uv);
 
 /*
  * Task to read the temperature from the battery pack
